requirements.txt
click>=8.0
pypdf>=3.0
python-docx>=0.8
python-pptx>=0.6
openpyxl>=3.1
pyyaml>=6.0
markdown>=3.4
